add immutable BSPNode; remove CSGSolid; add kdocs; avoid global namespace pollution
